refactor: rename bridge_test.go to rollup_test.go
